+++ pephub/main.py
@@ -1,25 +1,16 @@
 import sys
-import os
 import logmuse
 import uvicorn
+import os
 
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from starlette.staticfiles import StaticFiles
 
-from .pepstat import PEPIndexer
-from pephub.exceptions import PepHubException
-
-# set up global pep storage
-global _PEP_STORES  # the object in memory to read from
-global _PEP_STORAGE_PATH  # the actual file path to the peps
-_PEP_STORES = PEPIndexer()
-_PEP_STORAGE_PATH = ""
-
 from ._version import __version__ as server_v
 from .const import LOG_FORMAT, PKG_NAME, TAGS_METADATA
-from .helpers import build_parser, read_server_configuration
-from .routers import version1, namespace, project, eido, pep
+from .helpers import build_parser
+from .routers import version1, auth, namespace, project, eido, pep
 from .const import STATICS_PATH, EIDO_PATH
 
 # build server
@@ -42,6 +33,7 @@ app.add_middleware(
 
 # build routes
 app.include_router(version1.router)
+app.include_router(auth.router)
 app.include_router(namespace.router)
 app.include_router(project.router)
 app.include_router(eido.router)
@@ -56,45 +48,13 @@ app.mount(
 
 # The eido validator is an SPA that can be served as a static HTML
 # file. These can only be added on the main app, not on a router
-app.mount(
-    "/eido/validator", 
-    StaticFiles(directory=EIDO_PATH), 
-    name="eido_validator"
-)
-
-# populate config
-# read in the configration file
-cfg = read_server_configuration("config.yaml")
-
-# read in PEPs
-_PEP_STORAGE_PATH = cfg["data"]["path"]
-_INDEX_FILE = cfg["data"]["index"]
-
-if _INDEX_FILE is None:
-    print("No index file found. Indexing PEPs to 'index.yaml'")
-    _PEP_STORES.index(_PEP_STORAGE_PATH, "index.yaml")
-else:
-    if not os.path.exists(_INDEX_FILE):
-        print(f"Specified index file '{_INDEX_FILE}' does not exist. Creating index there.")
-        _PEP_STORES.index(_PEP_STORAGE_PATH, _INDEX_FILE)
-    else:
-        print(f"Loading index file: {_INDEX_FILE}")
-        _PEP_STORES.load_index(_INDEX_FILE)
+app.mount("/eido/validator", StaticFiles(directory=EIDO_PATH), name="eido_validator")
 
 def main():
     # set up the logger
     global _LOGGER
     parser = build_parser()
     args = parser.parse_args()
-
-    if args.config is None:
-        raise PepHubException(
-            "Configuration file required! " + "Please specify with '--config' flag."
-        )
-
-    # populate config
-    # read in the configration file
-    cfg = read_server_configuration(args.config)
 
     if not args.command:
         parser.print_help()