+++ pephub/routers/namespace.py
@@ -1,17 +1,16 @@
-from tkinter import PROJECTING
-from fastapi import APIRouter, Depends, Request
+from json import load
+import shutil
+import tempfile
+from typing import List
+from fastapi import APIRouter, Depends, Request, UploadFile, File, Form
 from fastapi.responses import JSONResponse, HTMLResponse
 from fastapi.templating import Jinja2Templates
 
-from pephub.const import BASE_TEMPLATES_PATH, INFO_KEY
-from peppy import __version__ as peppy_version
+from pephub.const import BASE_TEMPLATES_PATH
+from peppy import __version__ as peppy_version, Project
 from platform import python_version
 
-from pephub.pepstat.const import N_SAMPLES_KEY, PROJECTS_KEY
 from .._version import __version__ as pephub_version
-
-# fetch peps
-from ..main import _PEP_STORES
 
 # load dependencies
 from ..dependencies import *
@@ -19,42 +18,99 @@ from ..dependencies import *
 # examples
 from ..route_examples import example_namespace
 
+from dotenv import load_dotenv
+load_dotenv()
+
 router = APIRouter(
     prefix="/pep/{namespace}",
-    dependencies=[Depends(verify_namespace)],
     tags=["namespace"],
 )
 
 templates = Jinja2Templates(directory=BASE_TEMPLATES_PATH)
 
+
 @router.get("/", summary="Fetch details about a particular namespace.")
-async def get_namespace(namespace: str):
+async def get_namespace(
+    namespace: str,
+    db: Connection = Depends(get_db),
+):
     """Fetch namespace. Returns a JSON representation of the namespace."""
-    nspace = _PEP_STORES.get_namespace(
-        namespace.lower(), 
-        ignore_projects=True
-    )
+    nspace = db.get_namespace_info(namespace)
     return JSONResponse(content=nspace)
 
+
 @router.get("/projects", summary="Fetch all projects inside a particular namespace.")
-async def get_namespace_projects(namespace: str, limit: int = 100):
+async def get_namespace_projects(
+    namespace: str, db: Connection = Depends(get_db), limit: int = 100
+):
     """Fetch the projects for a particular namespace"""
-    projects = _PEP_STORES.get_projects(namespace)
+    projects = db.get_projects_in_namespace(namespace)
     if limit:
-        return JSONResponse(content={k: projects[k] for k in list(projects.keys())[:limit]})
+        return JSONResponse(content={p.name: p.to_dict() for p in projects[:limit]})
     else:
         return JSONResponse(content=projects)
 
-@router.get("/view", summary="View a visual summary of a particular namespace.", response_class=HTMLResponse)
-async def namespace_view(request: Request, namespace: str):
+@router.get(
+    "/view",
+    summary="View a visual summary of a particular namespace.",
+    response_class=HTMLResponse,
+)
+async def namespace_view(
+    request: Request, namespace: str, db: Connection = Depends(get_db), session_info: dict = Depends(read_session_info)
+):
     """Returns HTML response with a visual summary of the namespace."""
-    nspace = _PEP_STORES.get_namespace(namespace)
-    tot_samples = sum([nspace[PROJECTS_KEY][p][INFO_KEY][N_SAMPLES_KEY] for p in nspace['projects'] ])
-    return templates.TemplateResponse("namespace.html", {
-        'namespace': nspace,
-        'request': request,
-        'tot_samples': tot_samples,
-        'peppy_version': peppy_version,
-        'python_version': python_version(),
-        'pephub_version': pephub_version
-    })
+    nspace = db.get_namespace_info(namespace)
+    return templates.TemplateResponse(
+        "namespace.html",
+        {
+            "namespace": nspace,
+            "request": request,
+            "peppy_version": peppy_version,
+            "python_version": python_version(),
+            "pephub_version": pephub_version,
+            "logged_in": session_info is not None
+        },
+    )
+
+@router.post("/submit", summary="Submit a PEP to the current namespace")
+async def submit_pep(
+    request: Request,
+    namespace: str,
+    session_info: dict = Depends(read_session_info),
+    project_name: str = Form(...),
+    tag: str = Form(...),
+    config_file: UploadFile = File(...),
+    other_files: List[UploadFile] = File(...),
+    db: Connection = Depends(get_db),
+):
+    if session_info is None:
+        raise HTTPException(403, "Please log in to submit a PEP.")
+
+    # create temp dir that gets deleted
+    # after endpoint execution ends
+    with tempfile.TemporaryDirectory() as dirpath:
+        # save config file in tmpdir
+        with open(f"{dirpath}/{config_file.filename}", "wb") as cfg_fh:
+            shutil.copyfileobj(config_file.file, cfg_fh)
+
+        # save any other files the user might have supplied
+        for upload_file in other_files:
+            # open new file inside the tmpdir
+            with open(f"{dirpath}/{upload_file.filename}", "wb") as local_tmpf:
+                shutil.copyfileobj(upload_file.file, local_tmpf)
+
+        p = Project(f"{dirpath}/{config_file.filename}")
+        p.name = project_name
+        db.upload_project(p, namespace=namespace, name=project_name, tag=tag)
+        return templates.TemplateResponse(
+            "submission.html",
+            {
+                "request": request,
+                "namespace": namespace,
+                "project_name": project_name,
+                "proj": p.to_dict(),
+                "config_file": config_file.filename,
+                "other_files": [f.filename for f in other_files],
+                "tag": tag
+            }
+        )