+++ docs/changelog.md
@@ -2,6 +2,16 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.3.0] - 2022-09-07
+### Added
+- User authentication to submit PEPs
+- More thorough out `/view` endpoints
+- More PEPs 🎉
+
+### Changed
+- PEPhub is now backed by a postgres database
+- Utilizes `pepgbagent` to interface with database
+
 ## [0.2.0] - 2022-05-16
 ### Added
 - Better `/view` endpoints (switch from cards to tables)